refactor(nlp): processors updated
