+++ packages/botonic-nlp/src/tasks/text-classification/process/processor.ts
@@ -1,6 +1,6 @@
 import { tensor } from '@tensorflow/tfjs-node'
 
-import { Sample } from '../../../dataset/types'
+import { Sample } from '../../../dataset/input-parser'
 import { LabelEncoder } from '../../../encode/label-encoder'
 import { OneHotEncoder } from '../../../encode/one-hot-encoder'
 import { PADDING_TOKEN, UNKNOWN_TOKEN } from '../../../preprocess/constants'
